code cleanup and alignments

+++ src/Pages/InventoryDetail2/InventoryDetail2.js
@@ -3,6 +3,7 @@ import "./InventoryDetail2.css";
 import { Button } from "react-bootstrap";
 import useInventory from "../../hooks/useInventory";
 import { useNavigate } from "react-router-dom";
+import { toast } from "react-toastify";
 
 const InventoryDetail2 = ({ car }) => {
   const { supplier, quantity, description, price, image, _id } = car;
@@ -10,15 +11,17 @@ const InventoryDetail2 = ({ car }) => {
   const navigate = useNavigate();
 
   const handleDelete = (_id) => {
-    console.log(_id);
-    fetch(`https://tranquil-fortress-67244.herokuapp.com/inventory/${_id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        const remaining = cars.filter((car) => car._id !== _id);
-        setCars(remaining);
-      });
+    if (window.confirm("Do You Want To Delete")) {
+      fetch(`https://tranquil-fortress-67244.herokuapp.com/inventory/${_id}`, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          const remaining = cars.filter((car) => car._id !== _id);
+          setCars(remaining);
+          toast("Item deleted successfully. Please reload");
+        });
+    }
   };
 
   return (
@@ -32,7 +35,7 @@ const InventoryDetail2 = ({ car }) => {
         <p>Description : {description}</p>
       </div>
       <div className="row justify-content-evenly">
-      <Button
+        <Button
           onClick={() => navigate(`/updateinventorydetail/${_id}`)}
           className="mt-3"
           variant="light"
@@ -48,14 +51,14 @@ const InventoryDetail2 = ({ car }) => {
           Delete
         </Button>
       </div>
-        <Button
-          onClick={() => navigate(`/addmycars/${_id}`)}
-          className="mt-3"
-          variant="dark"
-          style={{ width: "200px", margin: '0 auto', display: 'block' }}
-        >
-          Add To My Car
-        </Button>
+      <Button
+        onClick={() => navigate(`/addmycars/${_id}`)}
+        className="mt-3"
+        variant="dark"
+        style={{ width: "200px", margin: "0 auto", display: "block" }}
+      >
+        Add To My Car
+      </Button>
     </div>
   );
 };
